feat: profile swagger 작성
- swagger 작동되도록 파서 구성
- 추후 실제 값에 맞추어 swagger line 일부 수정해야 함
- 가독성을 위한 routes 정리
- 기타 자잘한 typo
