+++ routes/common_routes.py
@@ -2,8 +2,8 @@
 
 from flask import Blueprint
 from flask_restx import Api
-from controllers.profile02_controller import ProfileController
 from routes.photo_routes import photo_ns
+from routes.profile_routes import profile_ns
 
 
 # Blueprint 생성
@@ -27,21 +27,4 @@ api.add_namespace(photo_ns)
 ##########################
 
 # 네임스페이스 생성
-profile_ns = api.namespace("profile", description="프로필 생성 API")
-
-# 파일 업로드 파서 설정
-profile_upload_parser = photo_ns.parser()
-profile_upload_parser.add_argument(
-    "responseList",
-    location="files",
-    type=dict,
-    required=True,
-    help="퀴즈 문항별 응답 값 (JSON)"
-)
-
-# Swagger에 파일 업로드 필드 추가
-# 실제 경로는 /api/v1/remove/bg
-@profile_ns.route("/generate/pf")
-@profile_ns.expect(profile_upload_parser)
-class ProfileResource(ProfileController):
-    pass
+api.add_namespace(profile_ns)
